Set up initial Tuhi integration
Add the tuhi.js files that talk to dbus and hook them up through electron and
react. We can only communicate between the render and main process over
IPC so we have an extra communication layer there.

Also, for conveniece, the debug() macro in the react bits uses that same IPC
to talk back to the render process, so we have all debug output on the console
as well.

environment variable DEBUG="tuhi*" must be set for debug() to work.